Create Helper Function to Aggregate Referral Data in Affiliates List
Related to #469
